videoStream/requirements.txt
pyserial~=3.4
pyaudio~=0.2.11
opencv-python~=4.5.3
